Move client extension functions into service catalog pkg
