--- pkg/broker/broker.go
@@ -1,18 +1,0 @@
-package broker
-
-import (
-	"fmt"
-
-	"github.com/kubernetes-incubator/service-catalog/pkg/apis/servicecatalog/v1beta1"
-	"github.com/kubernetes-incubator/service-catalog/pkg/client/clientset_generated/clientset"
-	"k8s.io/apimachinery/pkg/apis/meta/v1"
-)
-
-func retrieveByName(cl *clientset.Clientset, name string) (*v1beta1.ClusterServiceBroker, error) {
-	broker, err := cl.ServicecatalogV1beta1().ClusterServiceBrokers().Get(name, v1.GetOptions{})
-	if err != nil {
-		return nil, fmt.Errorf("unable to get broker '%s' (%s)", name, err)
-	}
-
-	return broker, nil
-}
