+++ pkg/broker/describe_cmd.go
@@ -5,6 +5,7 @@ import (
 
 	"github.com/Azure/service-catalog-cli/pkg/command"
 	"github.com/Azure/service-catalog-cli/pkg/output"
+	"github.com/Azure/service-catalog-cli/pkg/service-catalog/client"
 	"github.com/spf13/cobra"
 )
 
@@ -41,7 +42,7 @@ func (c *describeCmd) run(args []string) error {
 }
 
 func (c *describeCmd) describe(name string) error {
-	broker, err := retrieveByName(c.Client, name)
+	broker, err := client.RetrieveBroker(c.Client, name)
 	if err != nil {
 		return err
 	}
