+++ pkg/instance/describe_cmd.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/Azure/service-catalog-cli/pkg/command"
 	"github.com/Azure/service-catalog-cli/pkg/output"
-	"github.com/Azure/service-catalog-cli/pkg/traverse"
+	"github.com/Azure/service-catalog-cli/pkg/service-catalog/client"
 	"github.com/spf13/cobra"
 )
 
@@ -56,21 +56,21 @@ func (c *describeCmd) run(args []string) error {
 }
 
 func (c *describeCmd) describe(name string) error {
-	instance, err := retrieveByName(c.Client, c.ns, name)
+	instance, err := client.RetrieveInstance(c.Client, c.ns, name)
 	if err != nil {
 		return err
 	}
 
 	output.WriteInstanceDetails(c.Output, instance)
 
-	bindings, err := traverse.InstanceToBindings(c.Client, instance)
+	bindings, err := client.RetrieveBindingsByInstance(c.Client, instance)
 	if err != nil {
 		return err
 	}
 	output.WriteAssociatedBindings(c.Output, bindings)
 
 	if c.traverse {
-		class, plan, broker, err := traverse.InstanceParentHierarchy(c.Client, instance)
+		class, plan, broker, err := client.InstanceParentHierarchy(c.Client, instance)
 		if err != nil {
 			return fmt.Errorf("unable to traverse up the instance hierarchy (%s)", err)
 		}
