+++ pkg/instance/provision_cmd.go
@@ -6,6 +6,7 @@ import (
 	"github.com/Azure/service-catalog-cli/pkg/command"
 	"github.com/Azure/service-catalog-cli/pkg/output"
 	"github.com/Azure/service-catalog-cli/pkg/parameters"
+	"github.com/Azure/service-catalog-cli/pkg/service-catalog/client"
 	"github.com/spf13/cobra"
 )
 
@@ -67,7 +68,7 @@ func (c *provisonCmd) run(args []string) error {
 }
 
 func (c *provisonCmd) provision(name string, params map[string]string, secrets map[string]string) error {
-	instance, err := provision(c.Client, c.ns, name, c.className, c.planName, params, secrets)
+	instance, err := client.Provision(c.Client, c.ns, name, c.className, c.planName, params, secrets)
 	if err != nil {
 		return err
 	}
