+++ pkg/plan/describe_cmd.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/Azure/service-catalog-cli/pkg/command"
 	"github.com/Azure/service-catalog-cli/pkg/output"
-	"github.com/Azure/service-catalog-cli/pkg/traverse"
+	"github.com/Azure/service-catalog-cli/pkg/service-catalog/client"
 	"github.com/kubernetes-incubator/service-catalog/pkg/apis/servicecatalog/v1beta1"
 	"github.com/spf13/cobra"
 )
@@ -61,30 +61,30 @@ func (c *describeCmd) describe(key string) error {
 	var plan *v1beta1.ClusterServicePlan
 	var err error
 	if c.lookupByUUID {
-		plan, err = retrieveByUUID(c.Client, key)
+		plan, err = client.RetrievePlanByID(c.Client, key)
 	} else {
-		plan, err = RetrieveByName(c.Client, key)
+		plan, err = client.RetrievePlanByName(c.Client, key)
 	}
 	if err != nil {
 		return err
 	}
 
 	// Retrieve the class as well because plans don't have the external class name
-	class, err := traverse.PlanToClass(c.Client, plan)
+	class, err := client.RetrieveClassByPlan(c.Client, plan)
 	if err != nil {
 		return err
 	}
 
 	output.WritePlanDetails(c.Output, plan, class)
 
-	instances, err := traverse.PlanToInstances(c.Client, plan)
+	instances, err := client.RetrieveInstancesByPlan(c.Client, plan)
 	if err != nil {
 		return err
 	}
 	output.WriteAssociatedInstances(c.Output, instances)
 
 	if c.traverse {
-		broker, err := traverse.ClassToBroker(c.Client, class)
+		broker, err := client.RetrieveBrokerByClass(c.Client, class)
 		if err != nil {
 			return err
 		}
